+++ packages/vendure-order-client/test/dev-server.ts
@@ -1,5 +1,6 @@
 /* eslint-disable */
 import { DefaultLogger, LogLevel, mergeConfig } from '@vendure/core';
+import { MolliePlugin } from '@vendure/payments-plugin/package/mollie/mollie.plugin';
 import {
   createTestEnvironment,
   registerInitializer,
@@ -8,7 +9,7 @@ import {
   testConfig,
 } from '@vendure/testing';
 import { TestServer } from '@vendure/testing/lib/test-server';
-import { initialData } from './initial-data.js';
+import { initialData } from './initial-data';
 (async () => {
   let server: TestServer;
   let adminClient: SimpleGraphQLClient;
@@ -22,6 +23,12 @@ import { initialData } from './initial-data.js';
       adminApiPlayground: true,
       shopApiPlayground: true,
     },
+    plugins: [
+      MolliePlugin.init({
+        vendureHost: 'my-vendure.io',
+        useDynamicRedirectUrl: true,
+      }),
+    ],
   });
 
   ({ server, adminClient, shopClient } = createTestEnvironment(config));